alg/requirements.txt
# Core dependencies
fastapi>=0.115.12
uvicorn[standard]>=0.34.2

# Data processing
numpy==1.26.4
pandas==2.2.3
pyarrow>=15.0.0
nptdms>=1.4.0

# HTTP clients
httpx>=0.25.0
requests>=2.32.0

# Tools
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.0.0

# HTML to PDF dependencies
playwright>=1.40.0
PyPDF2>=3.0.0